init e2e tests
