+++ src/__tests__/utils/gql.ts
@@ -1,0 +1,10 @@
+export const gql = (
+  chunks: TemplateStringsArray,
+  ...variables: unknown[]
+): string => {
+  return chunks.reduce(
+    (acc, chunk, index) =>
+      `${acc}${chunk}${index in variables ? String(variables[index]) : ``}`,
+    ``,
+  );
+};
