+++ src/features/draft/commands/handlers/__tests__/api-create-revision.spec.ts
@@ -1,4 +1,5 @@
 import { CommandBus } from '@nestjs/cqrs';
+import {prepareBranch} from "src/__tests__/utils/prepareBranch";
 import {
   ApiCreateRevisionCommand,
   ApiCreateRevisionCommandReturnType,
@@ -7,7 +8,7 @@ import { PrismaService } from 'src/infrastructure/database/prisma.service';
 import {
   createMock,
   createTestingModule,
-  prepareBranch,
+
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import { EndpointNotificationService } from 'src/infrastructure/notification/endpoint-notification.service';
