+++ src/features/draft/commands/handlers/__tests__/create-row.spec.ts
@@ -1,13 +1,12 @@
 import { Prisma } from '@prisma/client';
 import { BadRequestException } from '@nestjs/common';
 import { CommandBus } from '@nestjs/cqrs';
+import {prepareBranch, PrepareBranchReturnType} from "src/__tests__/utils/prepareBranch";
 import { PrismaService } from 'src/infrastructure/database/prisma.service';
 import { TransactionPrismaService } from 'src/infrastructure/database/transaction-prisma.service';
 import {
   createMock,
   createTestingModule,
-  prepareBranch,
-  PrepareBranchReturnType,
   testSchema,
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import { CreateRowCommand } from 'src/features/draft/commands/impl/create-row.command';
