+++ src/features/draft/commands/handlers/__tests__/remove-table.spec.ts
@@ -1,8 +1,9 @@
 import { CommandBus } from '@nestjs/cqrs';
 import { nanoid } from 'nanoid';
+import {prepareBranch} from "src/__tests__/utils/prepareBranch";
 import {
   createTestingModule,
-  prepareBranch,
+
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import { RemoveTableCommand } from 'src/features/draft/commands/impl/remove-table.command';
 import { RemoveTableHandlerReturnType } from 'src/features/draft/commands/types/remove-table.handler.types';
