+++ src/features/draft/commands/handlers/__tests__/transactional/create-schema.spec.ts
@@ -1,8 +1,8 @@
 import { CommandBus } from '@nestjs/cqrs';
 import * as objectHash from 'object-hash';
+import {prepareBranch} from "src/__tests__/utils/prepareBranch";
 import {
   createTestingModule,
-  prepareBranch,
   testSchema,
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import {
