+++ src/features/draft/commands/handlers/__tests__/transactional/internal-update-row.spec.ts
@@ -1,4 +1,5 @@
 import { CommandBus } from '@nestjs/cqrs';
+import {prepareBranch} from "src/__tests__/utils/prepareBranch";
 import {
   InternalUpdateRowCommand,
   InternalUpdateRowCommandReturnType,
@@ -8,7 +9,6 @@ import { TransactionPrismaService } from 'src/infrastructure/database/transactio
 import {
   createMock,
   createTestingModule,
-  prepareBranch,
   testSchema,
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import { DraftTransactionalCommands } from 'src/features/draft/draft.transactional.commands';
