+++ src/features/draft/commands/handlers/__tests__/transactional/rename-schema.spec.ts
@@ -1,7 +1,8 @@
 import { CommandBus } from '@nestjs/cqrs';
+import {prepareBranch} from "src/__tests__/utils/prepareBranch";
 import {
   createTestingModule,
-  prepareBranch,
+
 } from 'src/features/draft/commands/handlers/__tests__/utils';
 import {
   RenameSchemaCommand,
